+++ src/app/modules/admin-core/admin-core.module.ts
@@ -2,11 +2,10 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { Interceptor } from './services/interceptor/interceptor';
-// import { DebounceDirective } from './directives/debounceDirective.directive';
 
 
 @NgModule({
-  // declarations: [DebounceDirective],
+  declarations: [],
   imports: [
     CommonModule,
     HttpClientModule
