+++ src/app/modules/admin-shared/admin-shared.module.ts
@@ -42,14 +42,15 @@ import { DynamicFormComponent } from './components/dynamic-form/dynamic-form.com
 import { TextAreaComponent } from './components/text-area/text-area.component';
 import { AutoCompleteComponent } from './components/auto-complete/auto-complete.component';
 import { ButtonLoadingComponent } from './button-loading/button-loading.component';
+import { DebounceDirective } from './directives/debounceDirective.directive';
 
 
 
 
 @NgModule({
   // ButtonComponent
-  declarations: [CommingSoonComponent, DynamicFieldDirective, UnauthorizedComponent, 
-    InputComponent, SelectComponent, RadiobuttonComponent, DynamicFormComponent,
+  declarations: [DebounceDirective,CommingSoonComponent, DynamicFieldDirective, UnauthorizedComponent, 
+    InputComponent, SelectComponent, RadiobuttonComponent, DynamicFormComponent, 
     CheckboxComponent,MultiSelectComponent, DateComponent, SpinnerComponent, PasswordComponent, ConfirmDialogComponent, TextAreaComponent, AutoCompleteComponent, ButtonLoadingComponent],
   imports: [
     CommonModule,
@@ -104,6 +105,7 @@ import { ButtonLoadingComponent } from './button-loading/button-loading.componen
     MatButtonToggleModule,
     DynamicFormComponent,
     ButtonLoadingComponent,
+    DebounceDirective,
     NgxMatSelectSearchModule],
   schemas: [NO_ERRORS_SCHEMA, CUSTOM_ELEMENTS_SCHEMA],
   entryComponents: [
