+++ src/app/modules/admin-shared/index.ts
@@ -5,3 +5,4 @@ export * from './components/dynamic-form/dynamic-form.component';
 export * from './components/dynamic-field/dynamic-field.directive';
 export * from './confirm-dialog/confirm-dialog.component';
 export * from './button-loading/button-loading.component';
+export * from './directives/debounceDirective.directive';
