+++ src/app/modules/organisations/organisations-list/organisations-list.component.html
@@ -19,7 +19,8 @@
       <div class="list-menu">
         <mat-form-field class="custom-form-field">
           <mat-label>Search for Organisation</mat-label>
-          <input matInput #searchInput>
+          <!-- <input matInput #searchInput> -->
+          <input  [(ngModel)] = "search" [appDebounceDirective] = "search" (debounceClick)="debounceMethod()" #searchInput matInput>
           <button mat-icon-button matPrefix>
             <mat-icon>search</mat-icon>
           </button>
