+++ src/app/modules/organisations/organisations-list/organisations-list.component.ts
@@ -30,6 +30,7 @@ export class OrganisationsListComponent implements OnInit {
   status: any = '';
   orgObject: any;
   assignedStatus: any;
+  search: any;
   queryParams = {
     page: 1,
     size: 10,
@@ -41,21 +42,21 @@ export class OrganisationsListComponent implements OnInit {
 
   ngOnInit() {
 
-    fromEvent(this.searchInput.nativeElement, 'keyup').pipe(
-      // get value
-      map((event: any) => {
-        return event.target.value;
-      })
-      // if character length greater then 2
-      , filter(res => res.length > 2 || res.length == 0)
-      // Time in milliseconds between key events
-      // , debounceTime(1000)
-      // If previous query is diffent from current
-      , distinctUntilChanged()
-      // subscription for response
-    ).subscribe((text: string) => {
-      this.getOrganisationList();
-    });
+    // fromEvent(this.searchInput.nativeElement, 'keyup').pipe(
+    //   // get value
+    //   map((event: any) => {
+    //     return event.target.value;
+    //   })
+    //   // if character length greater then 2
+    //   , filter(res => res.length > 2 || res.length == 0)
+    //   // Time in milliseconds between key events
+    //   // , debounceTime(1000)
+    //   // If previous query is diffent from current
+    //   , distinctUntilChanged()
+    //   // subscription for response
+    // ).subscribe((text: string) => {
+    //   this.getOrganisationList();
+    // });
     this.paginator.page.subscribe((page: PageEvent) => {
       this.queryParams.page = page.pageIndex + 1;
       this.queryParams.size = page.pageSize;
@@ -63,6 +64,10 @@ export class OrganisationsListComponent implements OnInit {
     });
     this.getOrganisationList();
     this.createOrganisationForm();
+  }
+
+  debounceMethod() {
+    this.getOrganisationList();
   }
 
   // get color based on the status
@@ -138,7 +143,7 @@ export class OrganisationsListComponent implements OnInit {
       confirmButtonText: "YES",
       cancelButtonText: "NO"
     }
-      const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
       width: "310px",
       height: "200px",
       data: confirmData
