+++ src/app/modules/upload-status/users-csv/users-csv.component.html
@@ -10,7 +10,8 @@
       <div class="list-menu" >
         <mat-form-field class="custom-form-field">
           <mat-label>Search </mat-label>
-          <input matInput #searchInput>
+          <!-- <input matInput #searchInput> -->
+          <input  [(ngModel)] = "search" [appDebounceDirective] = "search" (debounceClick)="debounceMethod()" #searchInput matInput>
           <button mat-icon-button matPrefix>
             <mat-icon>search</mat-icon>
           </button>
