+++ src/app/modules/upload-status/users-csv/users-csv.component.ts
@@ -24,6 +24,7 @@ export class UsersCsvComponent implements OnInit {
   initialType: any = '';
   status: any = '';
   columns: any;
+  search: any;
   statusList = [
     {
       label: 'All',
@@ -57,26 +58,30 @@ export class UsersCsvComponent implements OnInit {
 
   ngOnInit() {
 
-    fromEvent(this.searchInput.nativeElement, 'keyup').pipe(
-      // get value
-      map((event: any) => {
-        return event.target.value;
-      })
-      // if character length greater then 2
-      , filter(res => res.length > 2 || res.length == 0)
-      // Time in milliseconds between key events
-      , debounceTime(1000)
-      // If previous query is diffent from current
-      , distinctUntilChanged()
-      // subscription for response
-    ).subscribe((text: string) => {
-      this.bulkUploadList();
-    });
+    // fromEvent(this.searchInput.nativeElement, 'keyup').pipe(
+    //   // get value
+    //   map((event: any) => {
+    //     return event.target.value;
+    //   })
+    //   // if character length greater then 2
+    //   , filter(res => res.length > 2 || res.length == 0)
+    //   // Time in milliseconds between key events
+    //   , debounceTime(1000)
+    //   // If previous query is diffent from current
+    //   , distinctUntilChanged()
+    //   // subscription for response
+    // ).subscribe((text: string) => {
+    //   this.bulkUploadList();
+    // });
     this.paginator.page.subscribe((page: PageEvent) => {
       this.queryParams.page = page.pageIndex + 1;
       this.queryParams.size = page.pageSize;
       this.bulkUploadList();
     });
+    this.bulkUploadList();
+  }
+
+  debounceMethod() {
     this.bulkUploadList();
   }
